feat: :lock: Can't C.UD if you're not the owner

+++ php/Actions/Groups/DeleteGroup.php
@@ -3,11 +3,11 @@
 namespace Crisis\Actions\Groups;
 
 use Crisis\Models\Group;
-use Crisis\Actions\InvokableEMAction;
+use Crisis\Actions\ProtectedInvokableEMAction;
 use \Psr\Http\Message\ServerRequestInterface as Request;
 use \Psr\Http\Message\ResponseInterface as Response;
 
-class DeleteGroup extends InvokableEMAction
+class DeleteGroup extends ProtectedInvokableEMAction
 {
   public function handle(Request $request, Response $response, array $args): Response
   {
@@ -15,6 +15,12 @@ class DeleteGroup extends InvokableEMAction
     $group = $this->em
       ->getRepository(Group::class)
       ->find((int) $args['group_id']);
+
+    // Check authorisations
+    $jwtPayload = (new \PsrJwt\Helper\Request())->getTokenPayload($request, 'jwt');
+    if (!$this->checkUser((int) $jwtPayload['user_id'], (int) $args['user_id']) && !$this->checkUser((int) $jwtPayload['user_id'], $group->owner->id)) {
+      return $this->createResponse(['stauts' => 401, 'message' => 'Unauthorized'], 401);
+    }
 
     $this->em->remove($group);
     $this->em->flush();
