+++ php/Actions/Relations/DeleteRelation.php
@@ -3,11 +3,11 @@
 namespace Crisis\Actions\Relations;
 
 use Crisis\Models\Relation;
-use Crisis\Actions\InvokableEMAction;
+use Crisis\Actions\ProtectedInvokableEMAction;
 use \Psr\Http\Message\ServerRequestInterface as Request;
 use \Psr\Http\Message\ResponseInterface as Response;
 
-class DeleteRelation extends InvokableEMAction
+class DeleteRelationextends extends ProtectedInvokableEMAction
 {
   public function handle(Request $request, Response $response, array $args): Response
   {
@@ -15,6 +15,12 @@ class DeleteRelation extends InvokableEMAction
     $relation = $this->em
       ->getRepository(Relation::class)
       ->find((int) $args['relation_id']);
+
+    // Check authorisations
+    $jwtPayload = (new \PsrJwt\Helper\Request())->getTokenPayload($request, 'jwt');
+    if (!$this->checkUser((int) $jwtPayload['user_id'], $relation->sender->id)) {
+      return $this->createResponse(['stauts' => 401, 'message' => 'Unauthorized'], 401);
+    }
 
     $this->em->remove($relation);
     $this->em->flush();
